Add files via upload
This project contains some of the insights of the data which we got from the college department. It would help the department to see the courses which are full and highly demanded, the faculty which is critical and also see the other faculties which have the same qualification so that they could be assigned to the demanded courses.